backend/requirements.txt
# --- Core ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
pydantic-settings==2.5.2
pydantic[email]==2.9.0
python-multipart==0.0.9

# --- DB ---
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# --- Auth & security ---
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# --- Payments ---
stripe==10.12.0

# --- Object storage ---
boto3==1.35.24

# --- PDF generation ---
reportlab==4.2.2
svglib==1.5.1
Pillow==10.4.0

# --- HTTP / email ---
httpx==0.27.2

# --- Observability ---
structlog==24.4.0
sentry-sdk[fastapi]==2.14.0

# --- Rate limit ---
slowapi==0.1.9

# --- Misc ---
python-dotenv==1.0.1
